Plugged in a working example of React Router. Not convinced yet.
Also this has made me strip out all of the code for using React as a
RequireJS module, since it turns out that React plugins don't like
working that way, which is is a shame. So instead I've thrown all of
that away, and just dropped them in as script tags in the old fashioned
way that works.